+++ packages/abap-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sap-ux/abap-deploy-config-writer
 
+## 0.0.100
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/project-access@1.29.18
+    -   @sap-ux/system-access@0.5.33
+
+## 0.0.99
+
+### Patch Changes
+
+-   Updated dependencies [224494c]
+    -   @sap-ux/ui5-config@0.26.5
+    -   @sap-ux/project-access@1.29.17
+    -   @sap-ux/system-access@0.5.33
+
+## 0.0.98
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/ui5-config@0.26.4
+    -   @sap-ux/system-access@0.5.33
+    -   @sap-ux/project-access@1.29.16
+
 ## 0.0.97
 
 ### Patch Changes