+++ packages/annotation-generator/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/annotation-generator
 
+## 0.3.21
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/project-access@1.29.18
+    -   @sap-ux/fiori-annotation-api@0.4.21
+
+## 0.3.20
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.17
+-   @sap-ux/fiori-annotation-api@0.4.20
+
+## 0.3.19
+
+### Patch Changes
+
+-   @sap-ux/fiori-annotation-api@0.4.19
+-   @sap-ux/project-access@1.29.16
+
 ## 0.3.18
 
 ### Patch Changes