+++ packages/nodejs-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/nodejs-utils
 
+## 0.1.8
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/btp-utils@1.0.2
+
 ## 0.1.7
 
 ### Patch Changes