+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/project-access
 
+## 1.29.18
+
+### Patch Changes
+
+-   c3ebc82: fix: wrong convert preview-config prerequisites check for usage of cds-plugin-ui5
+
+## 1.29.17
+
+### Patch Changes
+
+-   Updated dependencies [224494c]
+    -   @sap-ux/ui5-config@0.26.5
+
+## 1.29.16
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/ui5-config@0.26.4
+
 ## 1.29.15
 
 ### Patch Changes