+++ packages/project-integrity/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @sap-ux/project-integrity
 
+## 0.1.16
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/project-access@1.29.18
+
+## 0.1.15
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.17
+
+## 0.1.14
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.16
+
+## 0.1.13
+
+### Patch Changes
+
+-   5148043: Fix to ensure writing of encoded integrity data
+
 ## 0.1.12
 
 ### Patch Changes