+++ packages/ui5-library-reference-sub-generator/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sap-ux/ui5-library-reference-sub-generator
 
+## 0.0.52
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/project-access@1.29.18
+    -   @sap-ux/fiori-generator-shared@0.9.11
+    -   @sap-ux/telemetry@0.5.72
+    -   @sap-ux/ui5-library-reference-inquirer@0.3.92
+    -   @sap-ux/ui5-library-reference-writer@0.1.57
+
+## 0.0.51
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.17
+-   @sap-ux/telemetry@0.5.71
+-   @sap-ux/ui5-library-reference-writer@0.1.56
+-   @sap-ux/fiori-generator-shared@0.9.10
+-   @sap-ux/ui5-library-reference-inquirer@0.3.91
+
+## 0.0.50
+
+### Patch Changes
+
+-   Updated dependencies [011c8c5]
+    -   @sap-ux/telemetry@0.5.70
+    -   @sap-ux/fiori-generator-shared@0.9.9
+    -   @sap-ux/ui5-library-reference-inquirer@0.3.90
+    -   @sap-ux/project-access@1.29.16
+    -   @sap-ux/ui5-library-reference-writer@0.1.55
+
 ## 0.0.49
 
 ### Patch Changes